DROOLS-7156 Recheck proper integration of Native Image tests
