+++ drools-drl-quarkus-extension/drools-drl-quarkus/pom.xml
@@ -20,6 +20,12 @@
         <dependency>
             <groupId>org.drools</groupId>
             <artifactId>drools-engine</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.drools</groupId>
+                    <artifactId>drools-wiring-dynamic</artifactId>
+                </exclusion>
+            </exclusions>
         </dependency>
 
         <!-- quarkus -->
